+++ craft_parts/lifecycle_manager.py
@@ -92,6 +92,9 @@ class LifecycleManager:
         sources defined on the host.
     :param build_environment: An iterable of environment variables in name=value format
         to be set during the build step.
+    :param stage_packages_slice_support: Whether Chisel slices can be declared in the
+        `stage-packages` key. Defaults to True for backward compatibility, but new apps
+        should set this to False and use the `stage-slices` key instead.
     :param custom_args: Any additional arguments that will be passed directly
         to callbacks.
     """
@@ -123,6 +126,7 @@ class LifecycleManager:
         usrmerged_by_default: bool = False,
         use_host_sources: bool = False,
         build_environment: Iterable[str] | None = None,
+        stage_packages_slice_support: bool = True,
         **custom_args: Any,  # custom passthrough args
     ) -> None:
         # pylint: disable=too-many-locals
@@ -165,6 +169,7 @@ class LifecycleManager:
             base_layer_dir=base_layer_dir,
             base_layer_hash=base_layer_hash,
             usrmerged_by_default=usrmerged_by_default,
+            stage_packages_slice_support=stage_packages_slice_support,
             **custom_args,
         )
 
@@ -174,7 +179,14 @@ class LifecycleManager:
 
         part_list: list[Part] = []
         for name, spec in parts_data.items():
-            part = _build_part(name, spec, project_dirs, strict_mode, partitions)
+            part = _build_part(
+                name,
+                spec,
+                project_dirs,
+                strict_mode,
+                partitions,
+                stage_packages_slice_support=stage_packages_slice_support,
+            )
             _validate_part_dependencies(part, parts_data)
             part_list.append(part)
 
@@ -342,11 +354,15 @@ def _build_part(
     project_dirs: ProjectDirs,
     strict_plugins: bool,  # noqa: FBT001
     partitions: list[str] | None,
+    *,
+    stage_packages_slice_support: bool = True,
 ) -> Part:
     """Create and populate a :class:`Part` object based on part specification data.
 
     :param spec: A dictionary containing the part specification.
     :param project_dirs: The project's work directories.
+    :param stage_packages_slice_support: Whether Chisel slices can be declared in the
+        `stage-packages` key.
 
     :return: A :class:`Part` object corresponding to the given part specification.
     """
@@ -393,6 +409,7 @@ def _build_part(
         project_dirs=project_dirs,
         plugin_properties=properties,
         partitions=partitions,
+        stage_packages_slice_support=stage_packages_slice_support,
     )