+++ tests/integration/sequencer/test_sequencer.py
@@ -37,6 +37,7 @@ _pull_state_foo = textwrap.dedent(
       source-commit: ''
       source-submodules: null
       stage-packages: []
+      stage-slices: []
       overlay-packages: []
       overlay-recommended-packages: []
     project_options:
@@ -75,6 +76,7 @@ _pull_state_bar = textwrap.dedent(
       source-commit: ''
       source-submodules: null
       stage-packages: []
+      stage-slices: []
       overlay-packages: []
       overlay-recommended-packages: []
     project_options: