+++ tests/integration/sequencer/test_sequencer_overlay.py
@@ -43,6 +43,7 @@ _pull_state_foo = textwrap.dedent(
       source-commit: ''
       source-submodules: null
       stage-packages: []
+      stage-slices: []
       overlay-packages: []
       overlay-recommended-packages: []
     project_options:
@@ -81,6 +82,7 @@ _pull_state_bar = textwrap.dedent(
       source-commit: ''
       source-submodules: null
       stage-packages: []
+      stage-slices: []
       overlay-packages: []
       overlay-recommended-packages: []
     project_options: